Modify the contact picker for import contact in FDN list
Modified the import contact menu in FDN list
on clicking on it will directly launch the list of contacts
instead of going to the system UI chooser which includes
a "Recent document" and "Downloads" option. The old behavior
required the user to tap on the "Contacts" option in the
hamburger menu.  The recent documents and downloads really
didn't make any sense to show.

Change-Id: Iaf97d8dbfe8d98492768eda3c345e5892b7443a8
CRs-Fixed: 2171251